refactor(store): minor PR-review polish on catalog admin
- Split StoreCatalogAdminViewModel and ProductInputModel into their own
  files; StoreIndexViewModel.cs keeps the index-page view models.
- Reorder GetAllProductsForYearAsync as OrderByDescending(IsActive)
  ThenBy(Name) for direct readability.
- Hoist the Spanish 21% VAT default into a named const
  (SpanishStandardVatRatePercent) at the top of StoreAdminController.
- Normalize the Save error path: InvalidOperationException now uses
  ModelState.AddModelError + return view, matching the surrounding
  ArgumentException catch (TempData wouldn't render without a redirect).
